move to separate class
